requirements.txt
gpyopt
keras-tcn
missingpy
pytorch-ignite
pytorch>=1.1.0
keras==2.3.1
numpy>=1.17.2
tensorflow==1.15.0
tqdm>=4.36.1
argparse>=1.1
scipy>=1.3.1
pandas>=0.25.1
scikit-learn>=0.21.3
texttable>=1.6.2
matplotlib>=3.1.1
fancyimpute>=0.5.4
logging>=0.5.1.2
tcn>=3.0.1
streamlit>=0.57.1